+++ lib/src/widgets/grid_view.dart
@@ -1,6 +1,7 @@
 import 'dart:developer' as developer;
 
 import 'package:fl_paging/src/datasource/data_source.dart';
+import 'package:fl_paging/src/widgets/builder.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart' as widgets;
 
@@ -24,8 +25,11 @@ class GridView<T> extends BaseWidget<T> {
         ErrorBuilder  errorBuilder,
       DataSource<T> pageDataSource})
       : super(
-            itemBuilder: itemBuilder, pageDataSource: pageDataSource, key: key,
-      emptyBuilder: emptyBuilder, loadingBuilder: loadingBuilder, errorBuilder: errorBuilder);
+            itemBuilder: itemBuilder,
+      emptyBuilder: emptyBuilder,
+      loadingBuilder: loadingBuilder,
+      errorBuilder: errorBuilder,
+      pageDataSource: pageDataSource, key: key);
 
   @override
   GridViewState<T> createState() => GridViewState<T>();
@@ -35,6 +39,10 @@ class GridViewState<T> extends State<GridView<T>> {
   static const TAG = 'GridView';
 
   PagingState<T> _pagingState = PagingState.loading();
+
+  void retry() {
+    _loadPage(isRefresh: false);
+  }
 
   Future _loadPage({bool isRefresh = false}) async {
     developer.log('_loadPage [isRefresh]: [$isRefresh]', name: TAG);
@@ -64,6 +72,11 @@ class GridViewState<T> extends State<GridView<T>> {
           });
         });
       } else {
+        if(_pagingState is PagingStateError<T>) {
+          setState(() {
+            PagingState.loading();
+          });;
+        }
         widget.pageDataSource.loadPage().then((value) {
           final oldState = (_pagingState as PagingStateData);
           setState(() {
@@ -104,7 +117,11 @@ class GridViewState<T> extends State<GridView<T>> {
   Widget build(BuildContext context) {
     return _pagingState.when((datas, isLoadMore, isEndList) {
       if (datas.length == 0) {
-        return widget.emptyBuilder(context);
+        if (widget.emptyBuilder != null) {
+          return widget.emptyBuilder(context);
+        } else {
+          return Container();
+        }
       } else {
         Widget child = widgets.SliverGrid(
           gridDelegate: widget.delegate,